major cleanup + refactoring
