+++ Teslanyse/App/Extensions.swift
@@ -26,7 +26,18 @@ extension View {
 }
 
 extension Double {
+
     func customNumberFormat(formatType: NumberFormatType = .dollar) -> String {
+        /**
+         Extends the `Double` type to include a method for formatting numbers into custom string representations based on specified formats. The method `customNumberFormat(formatType:)` supports various formats including currency (with denominations such as dollars, millions, billions), percentages, plain numbers with scale abbreviations (e.g., K, M, B), and units of power and energy (Watt, Kilowatt, Megawatt, and their hourly equivalents).
+
+         - Parameters:
+            - formatType: An optional parameter of type `NumberFormatType` with a default value of `.dollar`. This parameter determines the format in which the number will be returned. Supported formats are `.dollar` for currency, `.percent` for percentages, `.number` for scaled numbers without any unit, `.power` for power consumption units, and `.energy` for energy units.
+
+         - Returns: A `String` representing the formatted number based on the selected `formatType`. This includes adding appropriate suffixes (e.g., "$", "%", "K", "M", "B", "W", "KW", "MW", "Wh", "KWh", "MWh") and formatting the number with decimal places where appropriate. The method ensures that large numbers are presented in a human-readable format by scaling them down and appending relevant units or symbols.
+
+         This extension method leverages `NumberFormatter` to format numbers, enabling customization such as the use of grouping separators and setting maximum fraction digits according to the format type chosen. It simplifies displaying numbers in a user-friendly manner across various contexts, including financial amounts, percentages, quantities, and measurements of power or energy.
+         */
         let numberFormatter = NumberFormatter()
         numberFormatter.numberStyle = .decimal
         numberFormatter.usesGroupingSeparator = false
