+++ Teslanyse/App/SettingsClass.swift
